move hlg tests
